python2 unicode support changes.
Bit of a descrepency, it seems we write unicode when writing file
but return <str> when no output file or '-' is specified. Adjust
test so that we don't use r"blah blah blah" raw string format
as this will screw up when we use python3.

python 2
  dot type from file is unicode
  raw string is <str>

python3
  dot type from file is <str>
  raw string is <str>

This means for the test to pass running python2 we need to encode the
return from the file (e.g. unicde) to convert to <str> so we can compare.
With python3 we can't do that (no encode method for <str>).
We need the result and expected result types to match (in python2 & python3)