python/samba/netcmd: Fix possible unicode error #TODO Check this
This patch sortof replaces the change made with b86c29b1e6fb
for this file. This change seems strange that a string would be encoded
(e.g. converted to bytes) to be passed to an api that takes a string
paramater. That is change_password takes a 'string' as a password.
Alternatively change_password would take bytes and the previous change
would be correct